+++ ionize/Ion/acidity.py
@@ -1,8 +1,32 @@
 import warnings
-from math import log
-from ..constants import gas_constant, kelvin
+from math import log, sqrt
+from ..constants import gas_constant, kelvin, pitts
 import numpy as np
 
+
+def activity(self, valence=None, ionic_strength=None,
+                         temperature=None):
+    """Return activity coefficients of a charge state at ionic strength I."""
+    _, ionic_strength, temperature = \
+        self._resolve_context(None, ionic_strength, temperature)
+
+    if valence is None:
+        valence = self._valence_zero()
+    else:
+        valence = np.int_(valence)
+
+    # There are two coefficients that are used repeatedly.
+    # Specified in Bahga.
+    A = (self._solvent.debye_huckel(temperature)*sqrt(ionic_strength) /
+         (1. + pitts * sqrt(ionic_strength))
+         )
+    # TODO: check if this is right
+    B = 0.1*ionic_strength  # Matching STEEP implementation.
+
+    # Use them to calculate the activity coefficients.
+    gamma = 10**((valence**2)*(B-A))
+
+    return gamma
 
 def acidity(self, ionic_strength=None, temperature=None):
     """Return the effective Ka values for the ion.
